5543: Integrate API and widget settings modal in one controller
We move the modal creation from a service to a controller
