+++ public/angular/js/app.js
@@ -103,8 +103,6 @@ require('./services/tags');
 require('./controllers/application');
 require('./controllers/choose_password');
 require('./controllers/fetching');
-require('./controllers/email');
-require('./controllers/email_modal');
 require('./controllers/login');
 require('./controllers/navbar');
 require('./controllers/password_reset');
@@ -115,10 +113,10 @@ require('./controllers/reports/show');
 require('./controllers/sources/form_modal');
 require('./controllers/sources/index');
 require('./controllers/sources/show');
-require('./controllers/api/settings');
-require('./controllers/api/settings_modal');
-require('./controllers/widgets/settings');
-require('./controllers/widgets/settings_modal');
+require('./controllers/settings/settings');
+require('./controllers/settings/settings_modal');
+require('./controllers/settings/email');
+require('./controllers/settings/email_modal');
 require('./controllers/users/form_modal');
 require('./controllers/users/index');
 require('./controllers/users/profile');
