+++ public/angular/js/controllers/settings/email.js
