+++ public/angular/js/controllers/settings/email_modal.js
