+++ public/angular/js/services/api_modal.js
@@ -1,21 +1,21 @@
-angular.module('Aggie')
+// angular.module('Aggie')
 
-.factory('APISettingsModal', function($modal) {
+// .factory('APISettingsModal', function($modal) {
 
-  return {
-    create: function(api, settings) {
-      var modalInstance = $modal.open({
-        controller: 'APISettingsModalInstanceController',
-        templateUrl: 'templates/api-settings-modal.html',
-        resolve: {
-          api: function() {
-            return api;
-          },
-          settingsValues: function() {
-            return settings;
-          }
-        }
-      });
-    }
-  };
-});
+//   return {
+//     create: function(api, settings) {
+//       var modalInstance = $modal.open({
+//         controller: 'APISettingsModalInstanceController',
+//         templateUrl: 'templates/api-settings-modal.html',
+//         resolve: {
+//           api: function() {
+//             return api;
+//           },
+//           settingsValues: function() {
+//             return settings;
+//           }
+//         }
+//       });
+//     }
+//   };
+// });
