+++ public/angular/templates/api-settings-modal.html
@@ -3,30 +3,30 @@
     <div class="modal-header">
       <button class="close" ng-click="close()" aria-hidden="true">×</button>
       <h4 class="modal-title">
-        <div class="icon-left" translate="api_settings_modal.title" translate-value-api="{{ api }}"></div>
+        <div class="icon-left" translate="api_settings_modal.title" translate-value-api="{{ item }}"></div>
       </h4>
     </div>
     <div class="modal-body">
       <alert type="info" ng-show="success">{{ "api_settings_modal.success" | translate }}</alert>
       <alert type="danger" ng-show="failure">{{ message }}</alert>
 
-      <ol ng-if="api === 'twitter'">
+      <ol ng-if="item === 'twitter'">
         <li translate>api_settings_modal.twitter.p1</li>
         <li translate>api_settings_modal.twitter.p2</li>
       </ol>
-      <ol ng-if="api === 'facebook'">
+      <ol ng-if="item === 'facebook'">
         <li translate>api_settings_modal.facebook.p1</li>
         <li translate>api_settings_modal.facebook.p2</li>
         <li translate>api_settings_modal.facebook.p3</li>
         <li translate>api_settings_modal.facebook.p4</li>
       </ol>
-      <ol ng-if="api === 'elmo'">
+      <ol ng-if="item === 'elmo'">
         <li translate>api_settings_modal.elmo.p1</li>
         <li translate>api_settings_modal.elmo.p2</li>
         <li translate>api_settings_modal.elmo.p3</li>
         <li translate>api_settings_modal.elmo.p4</li>
       </ol>
-      <ol ng-if="api === 'gplaces'">
+      <ol ng-if="item === 'gplaces'">
         <li translate>api_settings_modal.gplaces.p1</li>
         <li translate>api_settings_modal.gplaces.p2</li>
       </ol>
@@ -44,16 +44,16 @@
       </div>
     </div>
     <div class="modal-footer">
-      <button class="btn btn-default"  ng-disabled="settingsForm.$invalid" ng-click="test(api, settings)" ng-if="api !== 'elmo' && api !== 'gplaces'">
+      <button class="btn btn-default"  ng-disabled="settingsForm.$invalid" ng-click="test(item, settings)" ng-if="item !== 'elmo' && item !== 'gplaces'">
         <span class="fa fa-refresh refresh-animate" ng-show="loading"></span>
         <span ng-hide="loading">Test</span>
         <span ng-show="loading">Testing</span></button>
-      <button class="btn btn-default" aggie-confirm="{{ api_settings_modal.confirm_delete | translate }}"
-              on-confirm="delete(api, settings)">
+      <button class="btn btn-default" aggie-confirm="api_settings_modal.confirm_delete"
+              on-confirm="delete(item, settings)">
         {{ 'Delete' | translate }}
       </button>
       <button class="btn btn-default" ng-click="close()" translate>Cancel</button>
-      <button class="btn btn-primary" ng-disabled="settingsForm.$invalid" ng-click="save(form, api, settings)" translate>
+      <button class="btn btn-primary" ng-disabled="settingsForm.$invalid" ng-click="save(form, item, settings)" translate>
         Save &amp; Close
       </button>
     </div>
